Move functions to graph.py
